requirements/requirements.txt
Django==1.9.3
Jinja2==2.8
MarkupSafe==0.23
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
binaryornot==0.4.0
celery==3.1.20
chardet==2.3.0
click==6.3
cookiecutter==1.3.0
defusedxml==0.4.1
django-allauth==0.24.1
django-braces==1.8.1
django-celery==3.1.17
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.1
django-floppyforms==1.6.1
django-model-utils==2.4
django-secure==1.0.1
future==0.15.2
kombu==3.0.33
oauthlib==1.0.3
psycopg2==2.6.1
pymongo==3.2.1
python3-openid==3.0.9
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
ruamel.base==1.0.0
ruamel.yaml==0.11.6
six==1.10.0
sqlparse==0.1.18
whichcraft==0.1.1
whitenoise==2.0.6
